refactor: reorganize project structure to adhere to principles defined in backend adr files
