+++ server/Application/Auth/UsosUserDto.cs
@@ -1,3 +1,3 @@
-﻿namespace Application.DTOs;
+﻿namespace Application.Auth;
 
 public record UsosUserDto(string Id);
