+++ server/Application/Features/Users/Commands/ProvisionUser/ProvisionUserCommand.cs
@@ -1,0 +1,3 @@
+﻿namespace Application.Features.Users.Commands.ProvisionUser;
+
+public record ProvisionUserCommand(string OauthToken, string OauthVerifier);
