+++ server/Application/Features/Users/Commands/ProvisionUser/UserDto.cs
@@ -1,4 +1,5 @@
 ﻿using System;
-namespace Application.DTOs;
+
+namespace Application.Features.Users.Commands.ProvisionUser;
 
 public record UserDto(string Id, string HashedUsosId, DateTimeOffset JoinedDate);
