+++ server/Presentation/Controllers/OpinionController.cs
@@ -1,8 +1,3 @@
-using System.Threading.Tasks;
-
-using Domain;
-
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Presentation.Controllers
